+++ mqtt/Client/Packets/PubCompPacket.cs
@@ -6,9 +6,9 @@ using System.Threading.Tasks;
 
 namespace Mqtt.Client.Packets
 {
-    public class PubCompPacket(int packetID)
+    public class PubCompPacket(ushort packetID)
     {
-        public int PacketID { get; } = packetID;
+        public ushort PacketID { get; } = packetID;
 
         public byte[] Encode()
         {
@@ -41,7 +41,7 @@ namespace Mqtt.Client.Packets
             int remainingLength = data[1];
 
             // Variable Header
-            int packetID = data[2] << 8 | data[3];
+            ushort packetID = (ushort)(data[2] << 8 | data[3]);
 
             return new PubCompPacket(packetID);
         }
